Expand Dark Theme
Expanded the dark theme to admin/edit.php
-Modified heading layout to fit with new theme
Pulled internal javascript from files/settings.php, admin/index.php, and
admin/edit.php
-Stored javascript in external files/edit.js and files/admin.js
Fixed bug on admin/edit.php where extra space was  at the end of title
URL when trim field left blank
Misc formating/spacing fixes
