+++ files/dbcheck.php
@@ -4,8 +4,6 @@
 		header('HTTP/1.0 403 Forbidden');
 		exit;
 	}
-	//The code on this page was borrowed and tweaked from: 
-	//https://github.com/rconfig/rconfig/blob/master/www/install/lib/ajaxHandlers/ajaxDbTests.php
 	require("password.php");
 	ini_set('display_errors', 0);
 	$server = $ini["host"];
